Removal of an unnecessary `elif` / `else` block after `continue`.
The use of 'else' or 'elif' becomes redundant and can be dropped if the last statement under the main 'if/elif' block is a 'continue' statement. In case of an 'elif' after 'continue', it can be written as a separate 'if' block. For 'else' blocks after 'continue', the statements can be moved out of the 'else' block.