Remove nnecessary `elif` / `else` block after ``raise`` and``continue`` (#73)
* Replace empty collections comp w/ unary operator

* Correct synthax error in the ternary operation

* peer-review: removing useless comparison to None

As said by @LeStarch, None is also falsey so no need to leave the check here

* Peer-review: correct incorrect changes

* Removal of unnecessary `else`/`elif` used after `raise`.

The 'raise' statement disrupts the control flow, as it allows to exit the block. It is recommended to check other conditions with another 'if' statement and to get rid of 'else' statements, which are useless.

* Removal of an unnecessary `elif` / `else` block after `continue`.

The use of 'else' or 'elif' becomes redundant and can be dropped if the last statement under the main 'if/elif' block is a 'continue' statement. In case of an 'elif' after 'continue', it can be written as a separate 'if' block. For 'else' blocks after 'continue', the statements can be moved out of the 'else' block.