fix cookiecutter new proj template
Cookiecutter new project template had "CMAKE_CURRENT_LIST_DIR" for
FPrime.cmake file. This works fine if fprime submodule is inside the
project folder, but won't work if the fprime repo is outside of the
project folder. With the new fix it does not matter where the fprime
folder is located as the path will be found from FPRIME_FRAMEWORK_PATH.